@@ -17,12 +17,13 @@
     "axios": "^1.7.7",
     "next": "^15.0.1",
     "next-auth": "^4.24.10",
-    "react": "18.2.0",
+    "react": "^18.3.1",
     "react-daum-postcode": "^3.1.3",
-    "react-dom": "18.2.0",
+    "react-dom": "^18.3.1",
     "react-hook-form": "^7.53.1",
     "react-icons": "^5.3.0",
-    "react-toastify": "^10.0.6"
+    "react-toastify": "^10.0.6",
+    "zustand": "^5.0.0"
   },
   "devDependencies": {
     "@types/node": "^22",
@@ -31,7 +32,6 @@
     "eslint-config-next": "15.0.1",
     "postcss": "^8.4.47",
     "prisma": "^5.21.1",
-    "recoil": "^0.7.7",
     "tailwindcss": "^3.4.14",
     "ts-node": "^10.9.2",
     "typescript": "^5"
