@@ -1,5 +1,7 @@
+'use client';
+
 import { signIn, useSession } from 'next-auth/react';
-import { useRouter } from 'next/router';
+import { useRouter } from 'next/navigation';
 import { useEffect } from 'react';
 
 import { AiOutlineGoogle } from 'react-icons/ai';
