@@ -3,7 +3,6 @@
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
 import { SessionProvider } from 'next-auth/react';
-import { RecoilRoot } from 'recoil';
 
 import { ToastContainer } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
@@ -17,19 +16,17 @@ interface Props {
 
 export const NextProvider = ({ children }: Props) => {
   return (
-    <RecoilRoot>
-      <QueryClientProvider client={queryClient}>
-        <SessionProvider>
-          {children}
-          <ToastContainer
-            autoClose={1000}
-            pauseOnFocusLoss={false}
-            pauseOnHover={false}
-          />
-          <ReactQueryDevtools />
-        </SessionProvider>
-      </QueryClientProvider>
-    </RecoilRoot>
+    <QueryClientProvider client={queryClient}>
+      <SessionProvider>
+        {children}
+        <ToastContainer
+          autoClose={1000}
+          pauseOnFocusLoss={false}
+          pauseOnHover={false}
+        />
+        <ReactQueryDevtools />
+      </SessionProvider>
+    </QueryClientProvider>
   );
 };
 
