@@ -1,90 +1,80 @@
 'use client';
 
-import { currentStoreState, locationState, mapState } from '@/atom';
 import { StoreType } from '@/interface';
+import {
+  useCurrentStore,
+  useLocationStore,
+  useMapStore,
+} from '@/zustand_store/store';
 import { useEffect, useCallback } from 'react';
-import { useRecoilValue, useSetRecoilState, useRecoilState } from 'recoil';
 
 interface MarkerProps {
   stores: StoreType[];
 }
 
 export default function Markers({ stores }: MarkerProps) {
-  const map = useRecoilValue(mapState);
-  const setCurrentStore = useSetRecoilState(currentStoreState);
-  const [location, setLocation] = useRecoilState(locationState);
+  const map = useMapStore((state) => state.map);
+  const setCurrentStore = useCurrentStore((state) => state.setCurrentStore);
+  const setLocation = useLocationStore((state) => state.setLocation);
 
   const loadKakaoMarkers = useCallback(() => {
-    if (map) {
-      //식당 데이터 마커 띄우기
-      stores?.map((store) => {
-        var imageSrc = store?.category
-            ? `/images/markers/${store?.category}.png`
-            : '/images/markers/default.png', //마커이미지의 주소입니다
-          imageSize = new window.kakao.maps.Size(40, 40), //마커이미지의 크기입니다
-          imageOption = { offset: new window.kakao.maps.Point(27, 69) }; //마커이미지의 옵션입니다. 마커의 좌표와 일치시킬 이미지 안에서의 좌표를 설정합니다.
+    if (!map) return;
 
-        //마커의 이미지정보를 가지고 있는 마커이미지를 생성합니다
-        var markerImage = new window.kakao.maps.MarkerImage(
-          imageSrc,
-          imageSize,
-          imageOption
-        );
+    stores.forEach((store) => {
+      const imageSrc = store.category
+        ? `/images/markers/${store.category}.png`
+        : `/images/markers/default.png`;
+      const imageSize = new window.kakao.maps.Size(40, 40);
+      const imageOption = { offset: new window.kakao.maps.Point(27, 69) };
 
-        //마커가 표시될 위치입니다
-        var markerPosition = new window.kakao.maps.LatLng(
-          store?.lat,
-          store?.lng
-        );
+      const markerImage = new window.kakao.maps.MarkerImage(
+        imageSrc,
+        imageSize,
+        imageOption
+      );
 
-        //마커를 생성합니다
-        var marker = new window.kakao.maps.Marker({
-          position: markerPosition,
-          image: markerImage, //마커이미지 설정
-        });
+      const markerPosition = new window.kakao.maps.LatLng(
+        store.lat ?? 0, // 기본값 설정
+        store.lng ?? 0 // 기본값 설정
+      );
 
-        //마커가 지도 위에 표시되도록 설정합니다
-        marker.setMap(map);
-
-        //마커 커서가 오버되었을 때 마커 위에 표시할 인포윈도우 생성
-        var content = `<div class="infowindow">${store?.name}</div>`; //인포윈도우에 표시될 내용
-
-        //커스텀 오버레이를 생성합니다
-        var customOverlay = new window.kakao.maps.CustomOverlay({
-          position: markerPosition,
-          content: content,
-          xAnchor: 0.6,
-          yAnchor: 0.91,
-        });
-
-        //마커에 마우스오버 이벤트를 등록합니다
-        window.kakao.maps.event.addListener(marker, 'mouseover', function () {
-          //마커에 마우스오버 이벤트가 발생하면 커스텀 오버레이를 마커위에 표시합니다
-          customOverlay.setMap(map);
-        });
-
-        //마커에 마우스아웃 이벤트를 등록합니다
-        window.kakao.maps.event.addListener(marker, 'mouseout', function () {
-          //마커에 마우스아웃 이벤트가 발생하면 커스텀 오버레이를 제거합니다
-          customOverlay.setMap(null);
-        });
-
-        //선택한 가게 저장
-        window.kakao.maps.event.addListener(marker, 'click', function () {
-          setCurrentStore(store);
-          setLocation({
-            ...location,
-            lat: store.lat,
-            lng: store.lng,
-          });
-        });
+      const marker = new window.kakao.maps.Marker({
+        position: markerPosition,
+        image: markerImage,
       });
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [map, stores]);
+
+      marker.setMap(map);
+
+      const content = `<div class="infowindow">${store.name}</div>`;
+      const customOverlay = new window.kakao.maps.CustomOverlay({
+        position: markerPosition,
+        content: content,
+        xAnchor: 0.6,
+        yAnchor: 0.91,
+      });
+
+      window.kakao.maps.event.addListener(marker, 'mouseover', () => {
+        customOverlay.setMap(map);
+      });
+
+      window.kakao.maps.event.addListener(marker, 'mouseout', () => {
+        customOverlay.setMap(null);
+      });
+
+      window.kakao.maps.event.addListener(marker, 'click', () => {
+        setCurrentStore(store);
+
+        // lat, lng가 undefined일 수 있으므로 기본값 사용
+        const lat = store.lat ?? 37.497625203; // 기본값 설정
+        const lng = store.lng ?? 127.03088379; // 기본값 설정
+        setLocation(lat, lng);
+      });
+    });
+  }, [map, stores, setCurrentStore, setLocation]);
 
   useEffect(() => {
     loadKakaoMarkers();
-  }, [loadKakaoMarkers, map]);
-  return <></>;
+  }, [loadKakaoMarkers]);
+
+  return null;
 }
